Split server-modifiers tests into a folder and drop its max-lines exemption
The admin-modifiers server test file had grown to 1,326 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split
it into test/lib/server-modifiers/ — grouped by what each part checks
(crud, revenue/edit, scope+links, and the child-only add-on guard), with
the shared createData/lastModifier helpers lifted into helpers.ts — so
every file sits under 400 lines, then removed the file's exemption entry.

Splitting the monolith brought it back under jscpd's per-file scan cap,
which surfaced three duplicated blocks that had been silently passing
inside the oversized file. Each is now a small local helper
(seedListingsScopedModifier, linkSoloAddOnToBookableAloneChild,
expectChildLinkSaved). All original tests carry over unchanged and pass;
jscpd is back to 0% duplication.
