+++ test/lib/server-modifiers/crud.test.ts
@@ -1,0 +1,351 @@
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { handleRequest } from "#routes";
+import { toMinorUnits } from "#shared/currency.ts";
+import {
+  adminFormPost,
+  adminGet,
+  awaitTestRequest,
+  createTestManagerSession,
+  describeWithEnv,
+  expectFlashRedirect,
+  expectHtmlResponse,
+  expectStatus,
+  followRedirectWithFlash,
+  testRequiresAuth,
+} from "#test-utils";
+import { createData, lastModifier } from "./helpers.ts";
+
+describeWithEnv("server (admin modifiers)", { db: true }, () => {
+  describe("GET /admin/modifiers", () => {
+    testRequiresAuth("/admin/modifiers");
+
+    test("accessible to managers", async () => {
+      const response = await awaitTestRequest("/admin/modifiers", {
+        cookie: await createTestManagerSession(),
+      });
+      expectStatus(200)(response);
+    });
+
+    test("shows empty list when no modifiers exist", async () => {
+      const response = await adminGet("/admin/modifiers");
+      await expectHtmlResponse(
+        response,
+        200,
+        "Modifiers",
+        "No modifiers configured",
+      );
+    });
+
+    test("lists modifiers with their rule summary", async () => {
+      await adminFormPost("/admin/modifiers", createData({ name: "Loyalty" }));
+      const response = await adminGet("/admin/modifiers");
+      await expectHtmlResponse(response, 200, "Loyalty", "Discount · 10%");
+    });
+  });
+
+  describe("GET /admin/modifiers/new", () => {
+    test("shows the create form", async () => {
+      const response = await adminGet("/admin/modifiers/new");
+      await expectHtmlResponse(response, 200, "Add Modifier", "Direction");
+    });
+  });
+
+  describe("POST /admin/modifiers", () => {
+    test("creates a percentage discount modifier", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData(),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers",
+        "Modifier created",
+        true,
+      )(response);
+      const modifier = await lastModifier();
+      expect(modifier.name).toBe("Early bird");
+      expect(modifier.calc_kind).toBe("percent");
+      expect(modifier.calc_value).toBe(10);
+      expect(modifier.direction).toBe("discount");
+    });
+
+    test("creates an active modifier when the toggle is checked", async () => {
+      await adminFormPost("/admin/modifiers", createData());
+      expect((await lastModifier()).active).toBe(true);
+    });
+
+    test("creates an inactive modifier when the toggle is cleared", async () => {
+      const data = createData();
+      // Omitting `active` mirrors an unchecked checkbox.
+      const { active: _omit, ...withoutActive } = data;
+      await adminFormPost("/admin/modifiers", withoutActive);
+      expect((await lastModifier()).active).toBe(false);
+    });
+
+    test("creates a fixed charge modifier", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({
+          calc_kind: "fixed",
+          calc_value: "5",
+          direction: "charge",
+          name: "Booking surcharge",
+        }),
+      );
+      const modifier = await lastModifier();
+      expect(modifier.calc_kind).toBe("fixed");
+      expect(modifier.calc_value).toBe(5);
+      expect(modifier.direction).toBe("charge");
+    });
+
+    test("accepts a fixed amount at the currency's precision", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_kind: "fixed", calc_value: "5.50" }),
+      );
+      expect((await lastModifier()).calc_value).toBe(5.5);
+    });
+
+    test("rejects a fixed amount with too many decimals for the currency", async () => {
+      // A fixed amount is a currency value (major units, converted at resolve).
+      // 10.005 has 3 decimals — invalid in GBP (2) — so it's rejected rather
+      // than silently rounded when the modifier is applied. A percentage or
+      // multiplier keeps its precision, so this only guards the fixed kind.
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_kind: "fixed", calc_value: "10.005" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Amount has more decimal places than your currency allows",
+        false,
+      )(response);
+    });
+
+    test("stores the minimum order in minor units", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ min_subtotal: "50" }),
+      );
+      expect((await lastModifier()).min_subtotal).toBe(toMinorUnits(50));
+    });
+
+    test("defaults the minimum order to zero when blank", async () => {
+      await adminFormPost("/admin/modifiers", createData());
+      expect((await lastModifier()).min_subtotal).toBe(0);
+    });
+
+    test("rejects a negative minimum order", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ min_subtotal: "-5" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Minimum order must be a valid amount for your currency",
+        false,
+      )(response);
+    });
+
+    test("rejects a minimum order with too many decimals for the currency", async () => {
+      // 10.005 has 3 decimals — invalid in GBP (2). Without validation this
+      // would be rounded by toMinorUnits at save; instead it's rejected.
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ min_subtotal: "10.005" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Minimum order must be a valid amount for your currency",
+        false,
+      )(response);
+    });
+
+    test("stores a stock limit", async () => {
+      await adminFormPost("/admin/modifiers", createData({ stock: "5" }));
+      expect((await lastModifier()).stock).toBe(5);
+    });
+
+    test("defaults stock to unlimited (null) when blank", async () => {
+      await adminFormPost("/admin/modifiers", createData());
+      expect((await lastModifier()).stock).toBeNull();
+    });
+
+    test("rejects a non-numeric value", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_value: "abc" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Enter a valid number",
+        false,
+      )(response);
+    });
+
+    test("rejects a percentage above 100", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_value: "150" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Percentage must be greater than 0 and at most 100",
+        false,
+      )(response);
+    });
+
+    test("rejects a zero percentage", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_value: "0" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Percentage must be greater than 0 and at most 100",
+        false,
+      )(response);
+    });
+
+    test("re-renders the create form with the validation error", async () => {
+      const { cookie, response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({
+          calc_kind: "fixed",
+          calc_value: "0",
+          direction: "charge",
+        }),
+      );
+      const page = await followRedirectWithFlash(
+        response,
+        (request) => handleRequest(request),
+        cookie,
+      );
+      await expectHtmlResponse(page, 200, "Amount must be greater than 0");
+    });
+
+    test("rejects a non-positive multiplier", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_kind: "multiply", calc_value: "0" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Multiplier must be greater than 0",
+        false,
+      )(response);
+    });
+
+    test("rejects a non-positive fixed amount", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_kind: "fixed", calc_value: "0" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Amount must be greater than 0",
+        false,
+      )(response);
+    });
+
+    test("rejects an unknown modifier type", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ calc_kind: "bogus" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Invalid modifier type",
+        false,
+      )(response);
+    });
+
+    test("rejects an unknown direction", async () => {
+      const { response } = await adminFormPost(
+        "/admin/modifiers",
+        createData({ direction: "sideways" }),
+      );
+      await expectFlashRedirect(
+        "/admin/modifiers/new",
+        "Invalid direction",
+        false,
+      )(response);
+    });
+  });
+
+  describe("GET /admin/modifiers/:id/edit", () => {
+    test("shows the edit form with current values", async () => {
+      await adminFormPost("/admin/modifiers", createData({ name: "Editable" }));
+      const { id } = await lastModifier();
+      const response = await adminGet(`/admin/modifiers/${id}/edit`);
+      await expectHtmlResponse(response, 200, "Edit Modifier", "Editable");
+    });
+
+    test("shows the minimum order in major units on the edit form", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ min_subtotal: "50" }),
+      );
+      const { id } = await lastModifier();
+      const response = await adminGet(`/admin/modifiers/${id}/edit`);
+      await expectHtmlResponse(response, 200, 'value="50"');
+    });
+
+    test("shows the stock limit on the edit form", async () => {
+      await adminFormPost("/admin/modifiers", createData({ stock: "7" }));
+      const { id } = await lastModifier();
+      const response = await adminGet(`/admin/modifiers/${id}/edit`);
+      await expectHtmlResponse(response, 200, 'value="7"');
+    });
+
+    test("returns 404 for a missing modifier", async () => {
+      const response = await adminGet("/admin/modifiers/999/edit");
+      expectStatus(404)(response);
+    });
+
+    test("shows the revenue-correction form with its ledger warning", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ name: "Surcharge" }),
+      );
+      const { id } = await lastModifier();
+      const response = await adminGet(`/admin/modifiers/${id}/edit`);
+      const html = await response.text();
+      expect(html).toContain("Adjust revenue");
+      expect(html).toContain(`action="/admin/modifiers/${id}/revenue"`);
+      expect(html).toContain('name="total_revenue"');
+      expect(html).toContain("correcting entry to the money ledger");
+    });
+
+    test("shows the owner-only modifier ledger section", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ name: "Helmet hire" }),
+      );
+      const { id } = await lastModifier();
+      const response = await adminGet(`/admin/modifiers/${id}/edit`);
+      const html = await response.text();
+      expect(html).toContain("Account statement");
+      expect(html).toContain("Add entry");
+      expect(html).toContain(
+        `/admin/ledger/modifier/${id}/add?return_url=%2Fadmin%2Fmodifiers%2F${id}%2Fedit`,
+      );
+    });
+
+    test("omits the ledger section for managers", async () => {
+      await adminFormPost(
+        "/admin/modifiers",
+        createData({ name: "Manager visible" }),
+      );
+      const { id } = await lastModifier();
+      const response = await awaitTestRequest(`/admin/modifiers/${id}/edit`, {
+        cookie: await createTestManagerSession("mgr-modifier-edit"),
+      });
+      const html = await response.text();
+      expect(response.status).toBe(200);
+      expect(html).toContain("Manager visible");
+      expect(html).not.toContain("Account statement");
+      expect(html).not.toContain(`/admin/ledger/modifier/${id}/add`);
+    });
+  });
+});
